refactor: NoteCard 컴포넌트를 entities/note로 이동
