+++ src/entities/note/ui/NoteCard.tsx
